Ensure deface is enabled during precompiling
If `config.deface.enabled = false` and you run `deface:precompile` in that environment (e.g. production as per the given example) then the views will not compile. Using an environment variable and setting it whilst running the rake task manually or during deploy will ensure the views are compiled properly.